+++ R/clean.R
@@ -34,6 +34,10 @@
 #' @param format a date format that will be passed on to \code{\link{format_datetime}}, see Details
 #' @param currency_symbol the currency symbol to use, which will be guessed based on the input and otherwise defaults to the current system locale setting (see \code{\link{Sys.localeconv}})
 #' @param guess_each logical to indicate whether all items of \code{x} should be guessed one by one, see Examples
+#' @param max_date date to indicate to maximum allowed of \code{x}, which defaults to today. This is to prevent that \code{clean_Date("23-03-47")} will return 23 March 2047 and instead returns 23 March 1947 with a warning.
+#' @param format character string giving a date-time format as used by \link[base]{strptime}. 
+#' 
+#' For \code{clean_Date(..., guess_each = TRUE)}, this can be a vector of values to be used for guessing, see Examples.
 #' @param ... for \code{clean_Date} and \code{clean_POSIXct}: other parameters passed on these functions
 #' @inheritParams base::as.POSIXct
 #' @details
@@ -88,6 +92,9 @@
 #' clean_Date("14526", "Excel")
 #' clean_Date(c("1 Oct 13", "October 1st 2012")) # could not be fitted in 1 format
 #' clean_Date(c("1 Oct 13", "October 1st 2012"), guess_each = TRUE)
+#' clean_Date(c("12-14-13", "1 Oct 2012"), 
+#'            guess_each = TRUE,
+#'            format = c("d mmm yyyy", "mm-yy-dd")) # only these formats will be tried
 #' 
 #' clean_POSIXct("Created log on 2020/02/11 11:23 by user Joe")
 #' clean_POSIXct("Created log on 2020.02.11 11:23 by user Joe", tz = "UTC")
@@ -323,41 +330,100 @@ clean_percentage <- function(x, remove = "[^0-9.,-]", fixed = FALSE) {
 
 #' @rdname clean
 #' @export
-clean_Date <- function(x, format = NULL, guess_each = FALSE, ...) {
+clean_Date <- function(x, format = NULL, guess_each = FALSE, max_date = Sys.Date(), ...) {
 
   if (is.Date(x)) {
     # could also be POSIX, or just Date
     return(as.Date(x))
   }
   
+  if (!is.Date(max_date) & !is.infinite(max_date)) {
+    max_date <- suppressWarnings(suppressMessages(clean_Date(max_date)))
+    if (is.na(max_date)) {
+      stop("`max_date` must be a date.")
+    }
+  }
+  
   if (is.double(x)) {
     x <- as.integer(x)
   }
-
-  if (!is.null(format)) {
+  
+  if (!is.null(format) & length(format) == 1) {
     if (tolower(format) == "excel") {
-      return(as.Date(as.numeric(x), origin = "1899-12-30"))
+      final_result <- as.Date(as.numeric(x), origin = "1899-12-30")
     } else {
-      return(as.Date(x = x, format = format_datetime(format), ...))
+      final_result <- as.Date(x = x, format = format_datetime(format), ...)
+    }
+  } else {
+    if (guess_each == FALSE) {
+      final_result <- guess_Date(x = x, throw_note = TRUE)
+    } else {
+      if (length(format) > 1) {
+        # checking date according to set vector of format options
+        x_coerced <- rep(NA_character_, length(x))
+        n_coerced <- 0
+        for (i in seq_len(length(format))) {
+          x_coerced[is.na(x_coerced)] <- tryCatch(as.Date(x = as.character(x[is.na(x_coerced)]),
+                                                          format = format_datetime(format[i])),
+                                                  warning = function(w) NA_character_,
+                                                  error = function(e) NA_character_)
+          message("(format '", format[i], "' used for ", length(x_coerced[!is.na(x_coerced)]) - n_coerced, " items)")
+          n_coerced <- length(x_coerced[!is.na(x_coerced)])
+        }
+        final_result <- as.Date(as.double(x_coerced), origin = "1970-01-01")
+      } else {
+        final_result <- as.Date(unname(sapply(x, guess_Date, throw_note = FALSE, format = format)), origin = "1970-01-01")
+      }
     }
   }
+  
+  tryCatch({
+    time_lt <- as.POSIXlt(final_result)
+    time_lt[final_result > max_date]$year <- time_lt[final_result > max_date]$year - 100
+    x <- as.Date(time_lt)
+    if (any(year(final_result) != year(x), na.rm = TRUE)) {
+      warning("Some years were decreased by 100 to not exceed today. Use clean_Date(..., max_date = Inf) to prevent this.")
+    }
+  },
+  error = function(e) x <<- final_result)
 
-  if (guess_each == FALSE) {
-    guess_Date(x = x, throw_note = TRUE)
-  } else {
-    as.Date(unname(sapply(x, guess_Date, throw_note = FALSE)), origin = "1970-01-01")
-  }
+  x
 }
 
 #' @rdname clean
 #' @export
-clean_POSIXct <- function(x, tz = "", remove = "[^.0-9 :/-]", fixed = FALSE, ...) {
+clean_POSIXct <- function(x, tz = "", remove = "[^.0-9 :/-]", fixed = FALSE, max_date = Sys.Date(), ...) {
+  if (is.Date(x)) {
+    # could also be POSIX, or just Date
+    return(as.POSIXct(x))
+  }
+  
+  if (!is.Date(max_date) & !is.infinite(max_date)) {
+    max_date <- suppressWarnings(suppressMessages(clean_Date(max_date)))
+    if (is.na(max_date)) {
+      stop("`max_date` must be a date.")
+    }
+  }
+  
   x <- trimws(gsub_warn_on_error(remove, "", x, ignore.case = TRUE, fixed = fixed))
   x <- gsub("[\\./]", "-", x)
-  as.POSIXct(x, tz = tz, ...)
+  
+  tryCatch({
+    time_ct <- tryCatch(as.POSIXct(x, tz = tz, ...),
+                        error = function(e) clean_Date(x))
+    time_lt <- as.POSIXlt(time_ct, tz = tz)
+    time_lt[time_ct > max_date]$year <- time_lt[time_ct > max_date]$year - 100
+    x <- as.POSIXct(time_lt, tz = tz, ...)
+    if (any(year(time_ct) != year(x), na.rm = TRUE)) {
+      warning("Some years were decreased by 100 to not exceed today. Use clean_POSIXct(..., max_date = Inf) to prevent this.")
+    }
+  },
+  error = function(e) x <<- time_ct)
+
+  as.POSIXct(x)
 }
 
-guess_Date <- function(x, throw_note = TRUE) {
+guess_Date <- function(x, throw_note = TRUE, format_options = NULL) {
   msg_clean_as <- function(format_set, sep = " ") {
     if (throw_note == TRUE) {
       if (tolower(format_set) == "excel") {
@@ -367,7 +433,7 @@ guess_Date <- function(x, throw_note = TRUE) {
       }
     }
   }
-  
+
   x_numeric <- suppressWarnings(as.numeric(x))
   if (all(x_numeric %in% c(as.integer(as.Date("1970-01-01") - as.Date("1899-12-30")):as.integer(Sys.Date() - as.Date("1899-12-30"))), na.rm = TRUE)) {
     # is Excel date